Tests
Tests para las tablas Features, Alerta, Device, Family, Group, LocationNames,Location Group, Location Priority, Tracking, WifiScan, RawSensorData, MessageGuess, Probabilities, Telemetria, Status, Message, Prediction y GoApiResponse

+++ favorita-service/src/main/java/net/pacificsoft/microservices/favorita/models/RawSensorData.java
@@ -36,14 +36,17 @@ public class RawSensorData implements Serializable{
 	@Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
 	private long id;
+        
         @Column(name = "epoch", nullable = false)
         private long epoch;
+        
         @Column(name = "temperature", nullable = false)
-        private float temperature;
+        private double temperature;
         
         @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm")
         @Column(name = "epochDateTime", nullable = false)
         private Date epochDateTime;
+        
         @Column(name = "rawData", nullable = false)
         private String rawData;
         
@@ -63,24 +66,24 @@ public class RawSensorData implements Serializable{
             mappedBy = "rawSensorData")
         private Set<SigfoxMessage> sigfoxMessages = new HashSet<>();
 
-        public RawSensorData(long epoch, float temperature, Date epochDateTime, String rawData) {
+        public RawSensorData(long epoch, double temperature, Date epochDateTime, String rawData) {
             this.epoch = epoch;
             this.temperature = temperature;
             this.epochDateTime = epochDateTime;
             this.rawData = rawData;
         }      
 
-    public RawSensorData(long id, long epoch, float temperature, Date epochDateTime, String rawData, Device device) {
-        this.id = id;
-        this.epoch = epoch;
-        this.temperature = temperature;
-        this.epochDateTime = epochDateTime;
-        this.rawData = rawData;
-        this.device = device;
-    }
+        public RawSensorData(long id, long epoch, double temperature, Date epochDateTime, String rawData, Device device) {
+            this.id = id;
+            this.epoch = epoch;
+            this.temperature = temperature;
+            this.epochDateTime = epochDateTime;
+            this.rawData = rawData;
+            this.device = device;
+        }
 
-    public RawSensorData() {
-    }
+        public RawSensorData() {
+        }
 
         
         public long getId() {
@@ -99,11 +102,11 @@ public class RawSensorData implements Serializable{
             this.epoch = epoch;
         }
 
-        public float getTemperature() {
+        public double getTemperature() {
             return temperature;
         }
 
-        public void setTemperature(float temperature) {
+        public void setTemperature(double temperature) {
             this.temperature = temperature;
         }
